feat: install multer, set up route and res.json
This was fairly easy with multer.
I installed it, the receiving of the file is easy.
As long as the form has the correct enctype, all you need to do is
to give the post route an extra argument called upload.single("whateverthenamefieldofthatformis")
and you can receive the file with req.file.